backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
typing-extensions>=4.12.2
pyswisseph>=2.10.3.2
sqlalchemy>=2.0.32
psycopg[binary]>=3.2.1
alembic>=1.13.2
redis>=5.0.8
openai>=1.44.0
python-dotenv>=1.0.1
orjson>=3.10.7
httpx>=0.27.0
tzdata>=2025.2
pytest>=8.3.2
